Tile names are announced instead of ids when possible for navigation and map viewing.
Files changed are pokemon.lua, navigation.lua and view.lua. Inpassible
tiles, tile names and "object tiles" were sepparated into their own file,
tiles.lua.

The tile_name function returns the name of a tile with a given id, based
on the tile_names table. If this can not be found, the solution for now is
to just return the id of the tile.

Object_tiles contains a list of tiles that should be added to the objects
list, such as the pc or trees. The object function in tiles.lua returns
the appropriate object table to add to the objects list.

Added small trees (or trees that headbutt can be used on) to objects list
in tiles.lua.

Added water to tile_names.